+++ modules/core/db/init/hsql/10.create-db.sql
@@ -14,6 +14,7 @@ create table DDCIL_LAUNCHER_CMD (
     GROUP_ID varchar(36),
     DESCRIPTION longvarchar,
     TYPE_ varchar(50) not null,
+    SHOW_IN_MAIN_MENU boolean,
     SHORTCUT varchar(255),
     --
     primary key (ID)
@@ -82,6 +83,7 @@ create table DDCIL_LAUNCHER_COMMAND_GROUP (
     --
     NAME varchar(255) not null,
     CODE varchar(255),
+    SHOW_IN_MAIN_MENU boolean,
     --
     primary key (ID)
 )^
@@ -125,3 +127,21 @@ create table DDCIL_INPUT_PARAMETER_TRANSLATION (
     primary key (ID)
 )^
 -- end DDCIL_INPUT_PARAMETER_TRANSLATION
+-- begin DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION
+create table DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION (
+    ID varchar(36) not null,
+    VERSION integer not null,
+    CREATE_TS timestamp,
+    CREATED_BY varchar(50),
+    UPDATE_TS timestamp,
+    UPDATED_BY varchar(50),
+    DELETE_TS timestamp,
+    DELETED_BY varchar(50),
+    --
+    LOCALE varchar(255) not null,
+    TEXT varchar(255) not null,
+    LAUNCHER_COMMAND_GROUP_ID varchar(36) not null,
+    --
+    primary key (ID)
+)^
+-- end DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION