+++ modules/core/db/init/mysql/10.create-db.sql
@@ -71,6 +71,7 @@ create table DDCIL_LAUNCHER_CMD (
     GROUP_ID varchar(32),
     DESCRIPTION longtext,
     TYPE_ varchar(50) not null,
+    SHOW_IN_MAIN_MENU boolean,
     SHORTCUT varchar(255),
     --
     primary key (ID)
@@ -89,6 +90,7 @@ create table DDCIL_LAUNCHER_COMMAND_GROUP (
     --
     NAME varchar(255) not null,
     CODE varchar(255),
+    SHOW_IN_MAIN_MENU boolean,
     --
     primary key (ID)
 )^
@@ -124,3 +126,21 @@ create table DDCIL_BEAN_LAUNCHER_COMMAND (
     primary key (ID)
 )^
 -- end DDCIL_BEAN_LAUNCHER_COMMAND
+-- begin DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION
+create table DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION (
+    ID varchar(32),
+    VERSION integer not null,
+    CREATE_TS datetime(3),
+    CREATED_BY varchar(50),
+    UPDATE_TS datetime(3),
+    UPDATED_BY varchar(50),
+    DELETE_TS datetime(3),
+    DELETED_BY varchar(50),
+    --
+    LOCALE varchar(255) not null,
+    TEXT varchar(255) not null,
+    LAUNCHER_COMMAND_GROUP_ID varchar(32) not null,
+    --
+    primary key (ID)
+)^
+-- end DDCIL_LAUNCHER_COMMAND_GROUP_TRANSLATION