+++ modules/core/db/update/hsql/20/200306-2-updateLauncherCmd-addShowInMainMenu.sql
@@ -1,0 +1,1 @@
+alter table DDCIL_LAUNCHER_CMD add column SHOW_IN_MAIN_MENU boolean ;