+++ modules/global/src/de/diedavids/cuba/instantlauncher/entity/LauncherCommand.java
@@ -10,6 +10,8 @@ import com.haulmont.cuba.core.global.DeletePolicy;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import java.util.List;
+import java.util.Locale;
+import java.util.Optional;
 
 @DiscriminatorValue("LAUNCHER")
 @DiscriminatorColumn(name = "DTYPE", discriminatorType = DiscriminatorType.STRING)
@@ -22,7 +24,7 @@ public class LauncherCommand extends StandardEntity {
     @Column(name = "NAME", nullable = false)
     protected String name;
 
-    @Lookup(type = LookupType.DROPDOWN, actions = {"lookup"})
+    @Lookup(type = LookupType.DROPDOWN, actions = "lookup")
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "GROUP_ID")
     protected LauncherCommandGroup group;
@@ -34,6 +36,9 @@ public class LauncherCommand extends StandardEntity {
     @NotNull
     @Column(name = "TYPE_", nullable = false)
     protected String type;
+
+    @Column(name = "SHOW_IN_MAIN_MENU")
+    protected Boolean showInMainMenu;
 
     @Composition
     @OnDelete(DeletePolicy.CASCADE)
@@ -47,6 +52,14 @@ public class LauncherCommand extends StandardEntity {
 
     @Column(name = "SHORTCUT")
     protected String shortcut;
+
+    public Boolean getShowInMainMenu() {
+        return showInMainMenu;
+    }
+
+    public void setShowInMainMenu(Boolean showInMainMenu) {
+        this.showInMainMenu = showInMainMenu;
+    }
 
     public List<InputParameter> getInputParameters() {
         return inputParameters;
@@ -107,6 +120,21 @@ public class LauncherCommand extends StandardEntity {
 
     public List<LauncherCommandTranslation> getTranslations() {
         return translations;
+    }
+
+
+    public String translationForLocale(Locale locale) {
+        return Optional.ofNullable(getTranslations())
+                .map(translations ->
+                        translations
+                                .stream()
+                                .filter(translation -> locale.equals(translation.getLocale()))
+                                .map(LauncherCommandTranslation::getText)
+                                .findFirst()
+                                .orElse(getName()))
+                .orElse(getName());
+
+
     }
 
 }