+++ modules/web/src/de/diedavids/cuba/instantlauncher/web/screens/main/InstantLauncherAppMainWindow.java
@@ -1,23 +1,31 @@
 package de.diedavids.cuba.instantlauncher.web.screens.main;
 
 import com.haulmont.cuba.gui.components.RootWindow;
+import com.haulmont.cuba.gui.components.mainwindow.AppMenu;
 import com.haulmont.cuba.web.app.mainwindow.AppMainWindow;
-import de.diedavids.cuba.instantlauncher.web.launcher.LauncherCommandShortcutInitializer;
+import de.diedavids.cuba.instantlauncher.web.launcher.LauncherCommandsInitializer;
+import de.diedavids.cuba.instantlauncher.web.launcher.menu.AppMenuAdapter;
 
 import javax.inject.Inject;
 
 public class InstantLauncherAppMainWindow extends AppMainWindow {
 
     @Inject
-    protected LauncherCommandShortcutInitializer launcherCommandShortcutInitializer;
+    protected LauncherCommandsInitializer launcherCommandShortcutInitializer;
 
+    @Inject
+    private AppMenu mainMenu;
 
     @Override
     public void ready() {
         super.ready();
 
-        launcherCommandShortcutInitializer.initInstantLauncherShortcuts(
+        launcherCommandShortcutInitializer.initKeyboardShortcuts(
                 (RootWindow) getFrameOwner().getWindow()
+        );
+
+        launcherCommandShortcutInitializer.initMenuLauncherCommands(
+                AppMenuAdapter.of(mainMenu)
         );
     }
 }